+++ test/unit/animation.js
@@ -175,4 +175,27 @@
 
   });
 
+  asyncTest( 'GH-142 Velocity kick after all images are loaded', function(){
+    /*
+      `velocity` value shouldn't be set on setup, but rather after all images are loaded
+    */
+    expect(3);
+    var
+      $reel= $('#image').reel({
+        velocity: 3
+      })
+
+    equal( $reel.reel('velocity'), 0, '`velocity` value at initialization (setup)');
+
+    $reel.bind('loaded.test', function(){
+      equal( $reel.reel('velocity'), 0, '`velocity` value right after loading is complete');
+    });
+
+    $(document).bind('opening.test', function(){
+      equal( $reel.reel('velocity'), 3, '`velocity` value after initialization (incl. preload) is complete');
+      start();
+    });
+
+  });
+
 })(jQuery);